DB 변경되어 엔티티 내용 수정 + [상담 예약] 페이 로직 처리, 상담 직접적인 에러  수정 중

+++ backend/src/main/java/com/study/spring/cnslInfo/entity/CnslInfo.java
@@ -6,7 +6,7 @@ import java.util.Date;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import com.study.spring.member.entity.Member;
+import com.study.spring.Member.entity.Member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -44,7 +44,7 @@ public class CnslInfo {
 	private String cnslTp;
 	
 	@Column(name="cnsl_price")
-	private Integer cnslPrice;
+	private Long cnslPrice;
 	
 	@Column(name="cnsl_rate")
 	private float cnslRate;
